requirements.txt
requests
python-dotenv
openai
playwright
selenium
fastapi
uvicorn[standard]
aiofiles
